server/requirements.txt
Flask==3.0.3
pymongo==4.7.3
google-generativeai==0.7.1
pydantic==2.7.4
python-dotenv==1.0.1  # For managing environment variables